change fft to use complex type
I thought I was being clever by using separate arrays for real
and imaginary parts. Now I realize that what you actually care
for is the phase and magnitude of the complex structure, and
that most of the time will be spent reading and processing
these values, and not copying memory. So yeah. God bless.