fix midi note reading
there was some issues with repeat notes and notes getting stuck.
this was resolved by fixing up the synth code a bit, and by
having a bigger buffer for midi notes. for now I moved the midi
notes to the processing code, but Im not sure if this is what I
want