use custom allocator
So I am not entirely sure on how I want to manage the core api.
I feel like introducing an allocator is a good idea, since some
of the utilities, like fft and delay, require allocating memory.
I think generally going for a RAII approach is fine here since
all resource aquisition is upfront, and also when we want to
initialize it. I think I will allow each of the "new_" functions
to allocate the memory they need for consistency (as opposed to
passing an &obj) to modify). yeah.